+++ lib/src/string_encryption.dart
@@ -17,98 +17,113 @@
 ///
 ///
 
-import 'package:flutter_string_encryption/flutter_string_encryption.dart'
-    show PlatformStringCryptor;
+// import 'package:flutter_string_encryption/flutter_string_encryption.dart'
+//     show PlatformStringCryptor;
 
+import 'package:encrypt/encrypt.dart';
+
+///
 class StringCrypt {
+  ///
   StringCrypt({
-    String key,
-    String password,
-    String salt,
+    String? password,
+    String? key,
   }) {
-    if (key != null && key.trim().isNotEmpty) {
-      _key = key;
+    //
+    if (password == null || password.trim().isEmpty) {
+      password = null;
     } else {
-      _keyFromPassword(password, salt).then((String key) {
-        _key = key;
-      });
+      password = password.trim().padLeft(32, '+');
     }
+
+    if (key == null || key.trim().isEmpty) {
+      key = null;
+    } else {
+      key = key.trim().padLeft(32, '+');
+    }
+
+    _iv = key == null ? IV.fromSecureRandom(32) : IV.fromUtf8(key);
+
+    final _key =
+        password == null ? Key.fromSecureRandom(32) : Key.fromUtf8(password);
+
+    _encrypter = Encrypter(AES(_key));
   }
-  String _key;
-  static PlatformStringCryptor _crypto = PlatformStringCryptor();
+  IV? _iv;
+  late Encrypter _encrypter;
 
-  Future<String> en(String data, [String key]) => encrypt(data, key);
+  ///
+  Future<String> en(String data, [String? key]) => encrypt(data, key);
 
-  Future<String> encrypt(String data, [String key]) async {
-    if (key != null) {
-      key = key.trim();
-      if (key.isEmpty) key = null;
+  ///
+  Future<String> encrypt(String data, [String? key]) async {
+    //
+    if (key == null || key.trim().isEmpty) {
+      key = null;
+    } else {
+      key = key.trim().padLeft(32, '+');
     }
+
+    final iv = key == null ? _iv : IV.fromUtf8(key);
+
     String encrypt;
     try {
-      encrypt = await _crypto.encrypt(data, key ??= _key);
+      encrypt = _encrypter.encrypt(data, iv: iv).base64;
     } catch (ex) {
-      encrypt = "";
+      encrypt = '';
       getError(ex);
     }
     return encrypt;
   }
 
-  Future<String> de(String data, [String key]) => decrypt(data, key);
+  ///
+  Future<String> de(String data, [String? key]) => decrypt(data, key);
 
-  Future<String> decrypt(String data, [String key]) async {
-    if (key != null) {
-      key = key.trim();
-      if (key.isEmpty) key = null;
+  ///
+  Future<String> decrypt(String data, [String? key]) async {
+    //
+    if (key == null || key.trim().isEmpty) {
+      key = null;
+    } else {
+      key = key.trim().padLeft(32, '+');
     }
+
+    final iv = key == null ? _iv : IV.fromUtf8(key);
+
     String decrypt;
     try {
-      decrypt = await _crypto.decrypt(data, key ??= _key);
+      decrypt = _encrypter.decrypt(Encrypted.fromBase64(data), iv: iv);
     } catch (ex) {
-      decrypt = "";
+      decrypt = '';
       getError(ex);
     }
     return decrypt;
   }
 
-  // You will need a key to decrypt things and so on.
-  static Future<String> generateRandomKey() => _crypto.generateRandomKey();
-
-  // Generates a salt to use with [generateKeyFromPassword]
-  static Future<String> generateSalt() => _crypto.generateSalt();
-
-  // Gets a key from the given [password] and [salt].
-  static Future<String> generateKeyFromPassword(String password, String salt) =>
-      _crypto.generateKeyFromPassword(password, salt);
-
-  Future<String> _keyFromPassword(String password, String salt) async {
-    if (password == null || password.trim().isEmpty) return "";
-    String _salt;
-    if (salt == null || salt.trim().isEmpty) {
-      _salt = await generateSalt();
-    } else {
-      _salt = salt.trim();
-    }
-    return await generateKeyFromPassword(password, _salt);
-  }
-
+  ///
   bool get hasError => _error != null;
 
+  ///
   bool get inError => _error != null;
-  Object _error;
+  Object? _error;
 
-  Exception getError([Object error]) {
+  ///
+  Exception? getError([Object? error]) {
     // Return the stored exception
-    Exception ex = _error;
+    Exception? ex;
+    if (_error != null) {
+      ex = _error as Exception;
+    }
     // Empty the stored exception
     if (error == null) {
       _error = null;
     } else {
-      if (error is! Exception) error = Exception(error.toString());
+      if (error is! Exception) {
+        error = Exception(error.toString());
+      }
       _error = error;
     }
     // Return the exception just past if any.
-    if (ex == null) ex = error;
-    return ex;
+    return ex ??= error as Exception;
   }
 }